[refac]: renamed spec files
