+++ utils/docs/specs/attributespec.py
@@ -4,7 +4,7 @@ from snakemd import Document  # type: ignore
 
 from utils.commons.logger import LOGGER
 from utils.docs.specs.abstract import ODDElement
-from utils.docs.specs.ns import NS_MAP
+from utils.docs.specs.namespaces import NS_MAP
 from utils.docs.specs.utils import split_tag_and_ns, translate_datatype
