+++ utils/docs/specs/namespaces.py
